Fix more cycles, move TaskContext type to types.tsx
