+++ packages/dara-core/js/types/index.tsx
@@ -22,6 +22,7 @@ import type {
     DerivedDataVariable,
     DerivedVariable,
     FilterQuery,
+    GlobalTaskContext,
     JsComponent,
     LayoutComponentProps,
     NormalizedPayload,
@@ -75,6 +76,7 @@ export type {
     BaseComponentProps,
     DataVariable,
     FilterQuery,
+    GlobalTaskContext,
     DataFrame,
     Pagination,
     ResolvedDataVariable,
